# Allow plugins to set environment variables in their caller's scope