--- a/crates/nu-plugin/src/util/mod.rs
+++ b/crates/nu-plugin/src/util/mod.rs
@@ -1,0 +1,3 @@
+mod mutable_cow;
+
+pub(crate) use mutable_cow::*;